requirements.txt
Django==3.1
django-extensions==3.1.0
Pillow==9.2.0
djangorestframework==3.12
drf-yasg==1.20.0
django-cors-headers==3.11.0